requirements.txt
python-telegram-bot>=20.0
openpyxl>=3.0
psycopg2-binary>=2.9
